+++ src/Open3D/Container/CMakeLists.txt
@@ -8,8 +8,8 @@ set (KERNEL_CUDA_SRC
 )
 
 set (CONTAINER_SRC
+    AdvancedIndexing.cpp
     Broadcast.cpp
-    Indexing.cpp
     MemoryManager.cpp
     MemoryManagerCPU.cpp
     MemoryManagerCUDA.cu
