+++ src/UnitTest/Container/Tensor.cpp
@@ -25,8 +25,8 @@
 // ----------------------------------------------------------------------------
 
 #include "Open3D/Container/Tensor.h"
+#include "Open3D/Container/AdvancedIndexing.h"
 #include "Open3D/Container/Dtype.h"
-#include "Open3D/Container/Indexing.h"
 #include "Open3D/Container/Kernel/Kernel.h"
 #include "Open3D/Container/MemoryManager.h"
 #include "Open3D/Container/SizeVector.h"
